small bug fixes

+++ EventStormingBoard.Server/Models/NoteSizeDto.cs
@@ -2,8 +2,8 @@
 {
     public sealed class NoteSizeDto
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
         public double Width { get; set; }
         public double Height { get; set; }
     }
